Properly calculates score based on efficiency.

+++ MGD1402/MGD1402/Classes/MainGameScene.h
@@ -23,7 +23,7 @@
 + (MainGameScene *)scene;
 - (id)init;
 - (void)getHighScores;
-- (void)updateHighScores:(int)newScore;
+- (void)updateHighScores:(float)newScore;
 
 // -----------------------------------------------------------------------
 @end
